Fixes to the curve gen and rotation calcs, updated memory offset address for PATH.BIN
Removed scaling in the curve calcs, so the points you pass in are in screen pixel units.
Adjusted ranges for the gui for points entry.
Added rendering of the Condensed curve offset from the uncondensed version.
Fixed rendering of the curves to properly get to the end of the path.
Creatted natvis for my PathPoint struct to make debugging easier.