Updated GeneratePath function
Changed it to use the closer of the two points along the curve that is closest to the MinDelta instead of always using the point that moved us past MinDelta, because that point could be significantly farther along.
Also only calculate the Direction when we are going to add the point to the path.