Renamed to SplinePathGenerator, more tweaks to the path rendering
Fixed the dumb bug in my previous change to use the closest point along curve to the MinDelta.
Fixed the rendering of the Condensed path to handle the case when the last entry has repeats and the end of it would not render because there was no position info for it.