Adding "projects" of multiple Splines.
There's now a project panel and menu. You can add/remove Splines and their paths will list in the panel, clicking on any of those paths will select it and load it into the canvas. You can drag and drop the paths to order them as you like (this will become important in the future when you can export a project to a BIN).
The project stuff is eliminated when in ROTATE_360 mode.